Script now takes advantage of Edge bonus...
Microsoft provides users who use Microsoft Edge with extra points. The script will now perform a single search on Microsoft Edge to take advantage of the daily 5 extra points you get for using Microsoft Edge. This part is only performed if it is on a Windows Machine with Edge, otherwise it will skip and proceed normally.